+++ gladanalysis/routes/api/v1/ms_router.py
@@ -8,51 +8,10 @@ import requests
 from . import endpoints
 from gladanalysis.services import GeostoreService
 from gladanalysis.services import DateService
+from gladanalysis.services import SqlService
 from gladanalysis.responders import ErrorResponder
 from gladanalysis.utils.http import request_to_microservice
 from gladanalysis.validators import validate_geostore, validate_period, validate_admin, validate_use, validate_wdpa
-
-def format_glad_sql(from_year, from_date, to_year, to_date, iso=None, state=None, dist=None):
-
-    select_sql = 'SELECT lat, long, confidence_text, country_iso, state_id, dist_id, year, julian_day '
-    count_sql = 'SELECT count(julian_day) '
-    from_sql = 'FROM {} '.format(os.getenv('GLAD_INDEX_ID'))
-    order_sql = 'ORDER BY year, julian_day'
-
-    if (int(from_year) < 2015 or int(to_year) > 2017):
-        return jsonify({'errors': [{
-            'status': '400',
-            'title': 'GLAD period must be between 2015 and 2017'
-            }]
-        }), 400
-
-    elif (from_year == '2015') and (to_year == '2017'):
-        where_template = ("WHERE ((year = '2015' and julian_day >= {d1}) or "
-        "(year = '2016') or "
-        "(year = '2017' and julian_day <= {d2}))")
-
-    elif from_year == to_year:
-        where_template = 'WHERE ((year = {Y1} and julian_day >= {d1} and julian_day <= {d2}))'
-
-    else:
-        where_template = 'WHERE ((year = {y1} and julian_day >= {d1}) or (year = {y2} and julian_day <= {d2}))'
-
-    geog_id_list = ['country_iso', 'state_id', 'dist_id']
-    geog_val_list = [iso, state, dist]
-
-    for geog_name, geog_value in zip(geog_id_list, geog_val_list):
-        if geog_value:
-            if geog_name == 'country_iso':
-                where_template += " AND ({} = '{}')".format(geog_name, geog_value)
-            else:
-                where_template += ' AND ({} = {})'.format(geog_name, geog_value)
-
-    where_sql = where_template.format(y1=from_year, d1=from_date, y2=to_year, d2=to_date)
-
-    sql = '?sql=' + ''.join([count_sql, from_sql, where_sql])
-    download_sql = '?sql=' + ''.join([select_sql, from_sql, where_sql, order_sql])
-
-    return sql, download_sql
 
 def format_terrai_sql(from_year, from_date, to_year, to_date, iso=None, state=None, dist=None):
 
@@ -160,7 +119,7 @@ def query_glad():
     from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
-    sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date)
+    sql, download_sql = SqlService.format_glad_sql(from_year, from_date, to_year, to_date)
 
     #create condition to look for confidence filter
     if conf == 'true' or conf == 'True':
@@ -223,7 +182,7 @@ def glad_dist(iso_code, admin_id, dist_id):
     from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
-    sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date, iso_code, admin_id, dist_id)
+    sql, download_sql = SqlService.format_glad_sql(from_year, from_date, to_year, to_date, iso_code, admin_id, dist_id)
 
     area_ha = GeostoreService.make_gadm_request(iso_code, admin_id)
 
@@ -255,7 +214,7 @@ def glad_admin(iso_code, admin_id):
     from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
-    sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date, iso_code, admin_id)
+    sql, download_sql = SqlService.format_glad_sql(from_year, from_date, to_year, to_date, iso_code, admin_id)
 
     #get geostore id from admin areas and total area of geostore request
     area_ha = GeostoreService.make_gadm_request(iso_code, admin_id)
@@ -288,7 +247,7 @@ def glad_country(iso_code):
     from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
-    sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date, iso_code)
+    sql, download_sql = SqlService.format_glad_sql(from_year, from_date, to_year, to_date, iso_code)
 
     #get geostore id from admin areas and total area of geostore request
     area_ha = GeostoreService.make_gadm_request(iso_code)
@@ -396,7 +355,7 @@ def glad_use(use_type, use_id):
     from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
-    sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date)
+    sql, download_sql = SqlService.format_glad_sql(from_year, from_date, to_year, to_date)
 
     geostore, area = GeostoreService.make_use_request(use_type, use_id)
 
@@ -452,7 +411,7 @@ def glad_wdpa(wdpa_id):
     from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
-    sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date)
+    sql, download_sql = SqlService.format_glad_sql(from_year, from_date, to_year, to_date)
 
     geostore, area = GeostoreService.make_wdpa_request(wdpa_id)
